# openapi.yaml
openapi: 3.0.1
info:
  title: Space NeoHoods API
  version: "2.0"
  description: |
    # Summary
    Space NeoHoods API
servers:
  - url: https://space.neohoods.com
    description: Space neohoods API
paths:
...
  /public/auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset password for a user
      tags:
        - auth-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "user@example.com"
              required:
                - email
      responses:
        "200":
          description: Password reset email sent
        "404":
          $ref: "#/components/responses/NotFound"
  /public/auth/reset-password/confirm:
    post:
      operationId: confirmResetPassword
      summary: Confirm password reset with token
      tags:
        - auth-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "reset-token-123"
                newPassword:
                  type: string
                  example: "newSecurePassword"
              required:
                - token
                - newPassword
      responses:
        "200":
          description: Password reset successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/auth/email-verification:
    post:
      operationId: verifyEmail
      summary: Verify user's email with token
      tags:
        - auth-api
      parameters:
        - name: token
          in: query
          description: Email verification token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email verification response
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Email verified successfully"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /admin/users/{user-id}/password:
    parameters:
      - name: user-id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
    put:
      operationId: setUserPassword
      summary: Set a user's password
      tags:
        - users-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "newSecurePassword"
              required:
                - newPassword
      responses:
        "204":
          description: Password updated successfully
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
...
  schemas:
...
    CodedError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A unique error code that identifies the type of error
          example: AUTH001
        message:
          type: string
          description: A human-readable error message
          example: The specified user could not be found
        traceId:
          type: string
          description: A unique identifier for tracing this error instance
          example: 550e8400-e29b-41d4-a716-446655440000
        documentationUrl:
          type: string
          description: URL to the documentation about this error
          example: https://docs.space.neohoods.com/errors/AUTH001
        variables:
          type: object
          additionalProperties: true
          description: Additional context variables for the error
          example:
            field: "email"
            value: "invalid@email"
...
  responses:
...
    BadRequest:
      description: Bad Request - The request was invalid or cannot be served
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          examples:
            validation_error:
              value:
                code: VAL001
                message: Validation failed
                variables:
                  fields:
                    - field: email
                      message: must be a valid email
            invalid_input:
              value:
                code: VAL002
                message: Missing required field
                variables:
                  field: username
...
    NotFound:
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          example:
            code: RES001
            message: The requested resource was not found
...